+++ src/properties.rs
@@ -57,24 +57,14 @@ impl<'a> Geo<'a> {
     where
         P: Into<Parameter<'a>>
     {
-        let parameter = parameter.into();
-        match self
-            .parameters
-            .iter_mut()
-            .find(|p| p.name == parameter.name)
-        {
-            Some(p) => *p = parameter,
-            None => self.parameters.push(parameter)
-        }
+        self.parameters.push(parameter.into())
     }
 
     /// Adds several parameters at once to the property. For creating
     /// several parameters at once, consult the documentation of
     /// the [`parameters!`] macro.
     pub fn append(&mut self, parameters: &mut Parameters<'a>) {
-        for parameter in parameters.drain(..) {
-            self.add(parameter);
-        }
+        self.parameters.append(parameters)
     }
 }
 
@@ -155,24 +145,14 @@ impl<'a> Status<'a> {
     where
         P: Into<Parameter<'a>>
     {
-        let parameter = parameter.into();
-        match self
-            .parameters
-            .iter_mut()
-            .find(|p| p.name == parameter.name)
-        {
-            Some(p) => *p = parameter,
-            None => self.parameters.push(parameter)
-        }
+        self.parameters.push(parameter.into())
     }
 
     /// Adds several parameters at once to the property. For creating
     /// several parameters at once, consult the documentation of
     /// the [`parameters!`] macro.
     pub fn append(&mut self, parameters: &mut Parameters<'a>) {
-        for parameter in parameters.drain(..) {
-            self.add(parameter);
-        }
+        self.parameters.append(parameters)
     }
 }
 
@@ -225,24 +205,14 @@ impl<'a> Transp<'a> {
     where
         P: Into<Parameter<'a>>
     {
-        let parameter = parameter.into();
-        match self
-            .parameters
-            .iter_mut()
-            .find(|p| p.name == parameter.name)
-        {
-            Some(p) => *p = parameter,
-            None => self.parameters.push(parameter)
-        }
+        self.parameters.push(parameter.into())
     }
 
     /// Adds several parameters at once to the property. For creating
     /// several parameters at once, consult the documentation of
     /// the [`parameters!`] macro.
     pub fn append(&mut self, parameters: &mut Parameters<'a>) {
-        for parameter in parameters.drain(..) {
-            self.add(parameter);
-        }
+        self.parameters.append(parameters)
     }
 }
 
@@ -386,24 +356,14 @@ mod rfc7986 {
         where
             P: Into<Parameter<'a>>
         {
-            let parameter = parameter.into();
-            match self
-                .parameters
-                .iter_mut()
-                .find(|p| p.name == parameter.name)
-            {
-                Some(p) => *p = parameter,
-                None => self.parameters.push(parameter)
-            }
+            self.parameters.push(parameter.into())
         }
 
         /// Adds several parameters at once to the property. For creating
         /// several parameters at once, consult the documentation of
         /// the [`parameters!`] macro.
         pub fn append(&mut self, parameters: &mut Parameters<'a>) {
-            for parameter in parameters.drain(..) {
-                self.add(parameter);
-            }
+            self.parameters.append(parameters)
         }
     }