+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Version 0.5.6
+
+### Bug Fix
+
+- Fix panic on folding multibytes character (see #15).
+
 ## Version 0.5.5
 
 ### Misc